[Chore] 프로젝트 모듈 분리 예시 작업
