+++ Projects/Cafe/Sources/APIClient.swift
@@ -7,6 +7,7 @@
 
 import Foundation
 import Dependencies
+import Network
 
 // TODO: 추후 API Sample code 정리 필요
 /// Sample Response
