+++ Projects/Cafe/Sources/Assets.xcassets/AccentColor.colorset/Contents.json
