+++ Projects/Cafe/Sources/Assets.xcassets/AppIcon.appiconset/Contents.json
