+++ Projects/Cafe/Sources/Assets.xcassets/Contents.json
