+++ Projects/Network/Project.swift
@@ -1,0 +1,16 @@
+import ProjectDescription
+import ProjectDescriptionHelpers
+
+private let projectName = "Network"
+private let iOSTargetVersion = "15.0"
+
+let project = Project.framework(
+    name: projectName,
+    product: .dynamicLibrary,
+    platform: .iOS,
+    iOSTargetVersion: iOSTargetVersion,
+    dependencies: [
+    ],
+    shouldIncludeTest: true,
+    shouldIncludeResources: true
+)
