+++ Projects/Network/Sources/CafeNetwork.swift
@@ -1,0 +1,13 @@
+//
+//  APIClient.swift
+//  YappProject
+//
+//  Created by Min Min on 2023/05/13.
+//
+
+public final class CafeNetwork {
+
+    public init() {
+        
+    }
+}
