--- Targets/22ndIOSTeam1IOS/Sources/AppDelegate.swift
@@ -1,25 +1,0 @@
-import UIKit
-import 22ndIOSTeam1IOSKit
-import 22ndIOSTeam1IOSUI
-
-@main
-class AppDelegate: UIResponder, UIApplicationDelegate {
-
-    var window: UIWindow?
-
-    func application(
-        _ application: UIApplication,
-        didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey : Any]? = nil
-    ) -> Bool {
-        window = UIWindow(frame: UIScreen.main.bounds)
-        let viewController = UIViewController()
-        viewController.view.backgroundColor = .white
-        window?.rootViewController = viewController
-        window?.makeKeyAndVisible()
-        22ndIOSTeam1IOSKit.hello()
-        22ndIOSTeam1IOSUI.hello()
-
-        return true
-    }
-
-}
