--- Targets/22ndIOSTeam1IOS/Tests/AppTests.swift
@@ -1,8 +1,0 @@
-import Foundation
-import XCTest
-
-final class 22ndIOSTeam1IOSTests: XCTestCase {
-    func test_twoPlusTwo_isFour() {
-        XCTAssertEqual(2+2, 4)
-    }
-}
