+++ Workspace.swift
@@ -1,0 +1,10 @@
+//
+//  Workspace.swift
+//  22nd-iOS-Team-1-iOSManifests
+//
+//  Created by 천수현 on 2023/05/21.
+//
+
+import ProjectDescription
+
+let workSpace = Workspace(name: "Cafe", projects: ["Projects/*"])
